requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.8.2
pydantic-settings==2.4.0

SQLAlchemy==2.0.32

openai==1.43.0
httpx==0.27.2
python-dotenv==1.0.1
